moved schema modules to folders
